+++ src/main/java/com/sesac/joinflex/domain/notification2/service/NotificationService.java
@@ -1,4 +1,4 @@
-package com.sesac.joinflex.domain.notification.service;
+package com.sesac.joinflex.domain.notification2.service;
 
 import java.io.IOException;
 import java.util.Map;
